requirements.txt
asgiref==3.2.3
coverage==5.0.3
Django==3.0.4
django-cors-headers==3.2.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
factory-boy==2.12.0
Faker==4.0.2
Pillow==7.0.0
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.3
